Welcome 2014!! from Curitiba back again!! Eu te amo Thomas Borregaard Sørensen!!
Eu te amo Mummi!! Feliz 62 anos!!

at inha at veriverse - at Dell XPS 8300 Intel(R) Core(TM) i7-2600 CPU 4 core 8 threads @ 3.40GHz 16GB RAM 500GB + 500GB 7200rpm storage + 120 GB SSD OCZ Agility 3 with Windows 8 Pro - with Portuguese - Brazil- 10Mpbs Download and 2Mpbs Upload NET Virtua Internet with LG 23´´ monitor, Microsoft Comfort Curve 2000 v1.0 Wired Keyboard and Microsoft Comfort Optical Wired Mouse 3000 at Alameda Princesa Izabel, 3000 - Curitiba - PR - Brazil

｛
｝

   ----
                        -   -   -
Mummi contribution!!, eu te amo demais!!
-----8nnnnnnnn8
++++++++++

+++ appseed/draw2d_gdi/draw2d_gdi_graphics.h
@@ -359,7 +359,7 @@ namespace draw2d_gdi
       virtual bool GetTextExtent(sized & size, const char * lpszString, strsize nCount, int32_t iIndex) const;
 
    // Text Functions
-      virtual bool TextOut(int x, int y, const char * lpszString, int nCount);
+      virtual bool TextOut(double x, double y, const char * lpszString, int nCount);
             bool TextOut(int x, int y, const string & str);
       virtual bool ExtTextOut(int x, int y, UINT nOptions, LPCRECT lpRect,
                const char * lpszString, UINT nCount, LPINT lpDxWidths);
